backend/pyproject.toml
[project]
name = "agent-ui-backend"
version = "0.1.0"
description = "Agent UI Backend with FastAPI and LangChain"
authors = [{ name = "Agent UI Team", email = "team@agentui.dev" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.104.1",
  "uvicorn[standard]>=0.24.0",
  "langchain>=0.0.340",
  "langchain-openai>=0.0.2",
  "langchain-community>=0.0.10",
  "pydantic>=2.5.0",
  "httpx>=0.25.2",
  "beautifulsoup4>=4.12.2",
  "python-dotenv>=1.0.0",
  "python-multipart>=0.0.6",
  "googlesearch-python>=1.2.3",
  "requests>=2.31.0",
  "aiofiles>=23.2.1",
  "python-json-logger>=2.0.7",
  "psutil>=5.9.6",
  "tavily-python>=0.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.3",
  "pytest-asyncio>=0.21.1",
  "pytest-cov>=4.1.0",
  "black>=23.11.0",
  "isort>=5.12.0",
  "flake8>=6.1.0",
  "mypy>=1.7.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["googlesearch.*", "beautifulsoup4.*", "bs4.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
